fix: fail if build fail

+++ tests/common.sh
@@ -1,7 +1,5 @@
 #/bin/sh
 
-set -e
-
 # Color codes.
 RED='\033[0;31m'
 GREEN='\033[0;32m'
@@ -28,7 +26,13 @@ run() {
     DOCKER_INTERACTIVE="-it"
     set -- sh
   fi
-  build >/dev/null
+  RET=$(build 2>&1 >/dev/null)
+  if  [[ $? -ne 0 ]]; then
+    echo "Build failed."
+    echo "${RET}"
+    exit 1
+  fi
+
 	docker run ${DOCKER_INTERACTIVE} --rm --entrypoint ash \
 		-v $(pwd)/build/docker-entrypoint.sh:/usr/local/bin/entrypoint.sh \
         -v $(pwd)/build/plugins.d:/usr/local/bin/plugins.d \
